Add a third database file describing database parameters.
The code in this commit works, but is superfluous, since the database
does not yet use it. The goal is to have the block size of .pag and .dir
files be configurable, using a ".par" file.

The .par file will be ASCII text, and have basic KEY=VALUE syntax. So far
the only keys are PBLKSIZ and DBLKSIZ. Their values must be powers of two
and may not exceed a value MAXBLKSIZ, defined in egdbm.h.

When creating a new database, default to PBLKSIZ and DBLKSIZ defined in
egdbm.h.  Perhaps a function like dbm_resize() can be implemented in the
future.